multi-agent-system/automation/requirements.txt
httpx==0.27.2
PyYAML==6.0.2
python-dotenv==1.0.1
